ads_celery/pyproject.toml
[tool.poetry]
name = "bp"
version = "0.1.0"
description = ""
authors = ["Real-4339 <qvadym@stuba.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
celery = "5.3.6"
loguru = "0.7.2"
redis = "5.0.4"
python-dotenv="1.0.1"
numpy = "2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

